Cleanup the EVM verifier (#195)
* WIP new serializer integration

* Finish integration

* Fixed assembly addresses offset bug

* Generate prover proof with new serializer

* Fix prover proof decoding case

* Fix optional field flags

* Fix t_comm missing commitments

* Remove commented line

* Refactor and simplify public commitment

* Remove unused functions

* Replace old ProverProof with new

* Remove LookupCommitments and some comments

* Remove serialization of old prover_proof, update kzg_prover test

* Remove duplicated code

* Resolve conflict

* Add Foundry installation step

* [WIP] Start fixing compiler errors

* Fix public comm usage

---------

+++ eth_verifier/lib/Constants.sol
@@ -1,31 +1,31 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.4.16 <0.9.0;
 
-uint constant COLUMNS = 15;
-uint constant PERMUTS = 7;
+uint256 constant COLUMNS = 15;
+uint256 constant PERMUTS = 7;
 
-uint constant PERMUTATION_CONSTRAINTS = 3;
-uint constant VARBASEMUL_CONSTRAINTS = 21;
+uint256 constant PERMUTATION_CONSTRAINTS = 3;
+uint256 constant VARBASEMUL_CONSTRAINTS = 21;
 
 // NewProofEvals bit flag positions
-uint constant PUBLIC_EVALS_FLAG = 0;
-uint constant RANGE_CHECK0_SELECTOR_EVAL_FLAG = 1;
-uint constant RANGE_CHECK1_SELECTOR_EVAL_FLAG = 2;
-uint constant FOREIGN_FIELD_ADD_SELECTOR_EVAL_FLAG = 3;
-uint constant FOREIGN_FIELD_MUL_SELECTOR_EVAL_FLAG = 4;
-uint constant XOR_SELECTOR_EVAL_FLAG = 5;
-uint constant ROT_SELECTOR_EVAL_FLAG = 6;
-uint constant LOOKUP_AGGREGATION_EVAL_FLAG = 7;
-uint constant LOOKUP_TABLE_EVAL_FLAG = 8;
-uint constant LOOKUP_SORTED_EVAL_FLAG = 9; // these are 5
-uint constant RUNTIME_LOOKUP_TABLE_EVAL_FLAG = 14;
-uint constant RUNTIME_LOOKUP_TABLE_SELECTOR_EVAL_FLAG = 15;
-uint constant XOR_LOOKUP_SELECTOR_EVAL_FLAG = 16;
-uint constant LOOKUP_GATE_LOOKUP_SELECTOR_EVAL_FLAG = 17;
-uint constant RANGE_CHECK_LOOKUP_SELECTOR_EVAL_FLAG = 18;
-uint constant FOREIGN_FIELD_MUL_LOOKUP_SELECTOR_EVAL_FLAG = 19;
+uint256 constant PUBLIC_EVALS_FLAG = 0;
+uint256 constant RANGE_CHECK0_SELECTOR_EVAL_FLAG = 1;
+uint256 constant RANGE_CHECK1_SELECTOR_EVAL_FLAG = 2;
+uint256 constant FOREIGN_FIELD_ADD_SELECTOR_EVAL_FLAG = 3;
+uint256 constant FOREIGN_FIELD_MUL_SELECTOR_EVAL_FLAG = 4;
+uint256 constant XOR_SELECTOR_EVAL_FLAG = 5;
+uint256 constant ROT_SELECTOR_EVAL_FLAG = 6;
+uint256 constant LOOKUP_AGGREGATION_EVAL_FLAG = 7;
+uint256 constant LOOKUP_TABLE_EVAL_FLAG = 8;
+uint256 constant LOOKUP_SORTED_EVAL_FLAG = 9; // these are 5
+uint256 constant RUNTIME_LOOKUP_TABLE_EVAL_FLAG = 14;
+uint256 constant RUNTIME_LOOKUP_TABLE_SELECTOR_EVAL_FLAG = 15;
+uint256 constant XOR_LOOKUP_SELECTOR_EVAL_FLAG = 16;
+uint256 constant LOOKUP_GATE_LOOKUP_SELECTOR_EVAL_FLAG = 17;
+uint256 constant RANGE_CHECK_LOOKUP_SELECTOR_EVAL_FLAG = 18;
+uint256 constant FOREIGN_FIELD_MUL_LOOKUP_SELECTOR_EVAL_FLAG = 19;
 
-// NewProofCommitments bit flag positions
-uint constant LOOKUP_SORTED_COMM_FLAG = 0;
-uint constant LOOKUP_AGGREG_COMM_FLAG = 1;
-uint constant LOOKUP_RUNTIME_COMM_FLAG = 2;
+// ProofCommitments bit flag positions
+uint256 constant LOOKUP_SORTED_COMM_FLAG = 0;
+uint256 constant LOOKUP_AGGREG_COMM_FLAG = 1;
+uint256 constant LOOKUP_RUNTIME_COMM_FLAG = 2;
